# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: LaunchPad API
  version: 0.8.0
servers:
  - url: http://localhost:8080/api/v1
paths:
...
  /journeys:
    get:
      summary: List journey templates
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
    post:
      summary: Create draft journey template
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Created
  /journeys/{journeyID}:
    get:
      summary: Get journey template
      security:
        - bearerAuth: []
      parameters:
        - name: journeyID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...